Tests/LibWeb/Text/expected/select-text-before-contenteditable.txt

selected post text: true
<BODY>
